back-end/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
Deprecated==1.2.14
Django==5.0.7
django-axes==6.5.1
django-allauth==0.52.0
django-guest-user
dnspython==2.6.1
Faker==26.0.0
faker-commerce==1.0.4
flasgger==0.9.7.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
gunicorn==22.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
mistune==3.0.2
mysql-connector-python==9.0.0
ordered-set==4.1.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
PyMySQL==1.1.1
pytest==8.2.2
python-decouple==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.28.1
rich==13.7.1
rpds-py==0.19.0
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
typing_extensions==4.12.2
Werkzeug==3.0.3
wrapt==1.16.0
